FEAT: refatora campos numéricos nas atividades para garantir valores inteiros
Atualiza a lógica de criação e atualização de atividades para utilizar a função `inteiroOuZero`, assegurando que campos numéricos sejam convertidos corretamente para inteiros ou zero. Implementa normalização de campos numéricos na interface do usuário e adiciona testes para validar o comportamento ao lidar com valores nulos e vazios, garantindo a integridade dos dados nas atividades.

+++ resources/views/livewire/admin/projeto-edit-drawer.blade.php
@@ -298,13 +298,13 @@
                                                             <div>
                                                                 <p class="text-xs text-default-600">Total Extensão</p>
                                                                 <p class="text-sm font-semibold text-primary">
-                                                                    {{ ($atividade['extensao_desenho'] ?? 0) + ($atividade['extensao_projeto'] ?? 0) }} m
+                                                                    {{ (int) ($atividade['extensao_desenho'] ?? 0) + (int) ($atividade['extensao_projeto'] ?? 0) }} m
                                                                 </p>
                                                             </div>
                                                             <div>
                                                                 <p class="text-xs text-default-600">Total Postes</p>
                                                                 <p class="text-sm font-semibold text-primary">
-                                                                    {{ ($atividade['postes_desenhados'] ?? 0) + ($atividade['postes_projetados'] ?? 0) }}
+                                                                    {{ (int) ($atividade['postes_desenhados'] ?? 0) + (int) ($atividade['postes_projetados'] ?? 0) }}
                                                                 </p>
                                                             </div>
                                                         </div>
